+++ example/test2.m
@@ -1,7 +1,1 @@
-configure('real', 'number')
-clear
-getconfig()
-e^(i*pi)
-e^(i*pi)==-1
-sin(pi/2)
-abs(-5)
+clear % clear all variables