+++ src/components/components.ts
@@ -1,4 +1,4 @@
-/* Web components. File generated at Sun, 26 Oct 2025 23:58:05 GMT. */
+/* Web components. File generated at Sun, 30 Nov 2025 17:50:47 GMT. */
 export * from './collapsible-content-panel/collapsible-content-panel.component';
 export * from './fixed-scroll-panel/fixed-scroll-panel.component';
 export * from './command-prompt/command-prompt.component';